offers: import latest variant from draft.
In particular, this changes the name of a field in invoice_request:
`payer_signature` becomes simply `signature`.  So we allow both for
now, and send the old one unless deprecated_apis is disabled.

+++ plugins/offers_invreq_hook.h
@@ -3,7 +3,7 @@
 #include "config.h"
 #include <plugins/libplugin.h>
 
-extern u32 cltv_final;
+extern u16 cltv_final;
 
 /* We got an onionmessage with an invreq! */
 struct command_result *handle_invoice_request(struct command *cmd,
